snapping stacks, yo
