+++ resources/baccarat-layout.json
@@ -17,7 +17,7 @@
          "name":"Object Layer 1",
          "objects":[
                 {
-                 "height":192,
+                 "height":224,
                  "id":1,
                  "name":"",
                  "properties":
@@ -27,12 +27,12 @@
                  "rotation":0,
                  "type":"player_bet",
                  "visible":true,
-                 "width":432,
-                 "x":496,
-                 "y":704
+                 "width":512,
+                 "x":448,
+                 "y":688
                 }, 
                 {
-                 "height":144,
+                 "height":160,
                  "id":2,
                  "name":"",
                  "properties":
@@ -42,12 +42,12 @@
                  "rotation":0,
                  "type":"dealer_bet",
                  "visible":true,
-                 "width":400,
-                 "x":512,
+                 "width":512,
+                 "x":448,
                  "y":528
                 }, 
                 {
-                 "height":112,
+                 "height":128,
                  "id":3,
                  "name":"",
                  "properties":
@@ -57,8 +57,8 @@
                  "rotation":0,
                  "type":"tie_bet",
                  "visible":true,
-                 "width":336,
-                 "x":544,
+                 "width":512,
+                 "x":448,
                  "y":400
                 }, 
                 {
@@ -102,9 +102,9 @@
                  "rotation":0,
                  "type":"player_chips",
                  "visible":true,
-                 "width":464,
-                 "x":48,
-                 "y":528
+                 "width":400,
+                 "x":32,
+                 "y":544
                 }, 
                 {
                  "height":144,
